copysignf

+++ paddle/phi/kernels/impl/copysign_kernel_impl.h
@@ -26,7 +26,7 @@ struct CopySignFunctor {
       : x_data_(x_data), y_data_(y_data), out_data_(out_data), numel_(numel) {}
 
   HOSTDEVICE void operator()(int64_t idx) const {
-    out_data_[idx] = static_cast<T>(std::copysign(
+    out_data_[idx] = static_cast<T>(std::copysignf(
         static_cast<T>(x_data_[idx]), static_cast<T>(y_data_[idx])));
   }
